A major overhaul
Added a Compiler so you can write the code in a code editor, then it adds the Universal Threading Library, then turns it into a vrPython file you can load.

! WORK IN PROGRESS
I also started a new feture so that you can run the python in your favored code editor, and it shows a turtle recreation of what would happen.